requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
_pytorch_select=0.2=gpu_0
absl-py=1.2.0=pyhd8ed1ab_0
aiohttp=3.8.1=py37h540881e_1
aiosignal=1.2.0=pyhd8ed1ab_0
anyio=3.5.0=py37h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py37h7f8727e_0
async-timeout=4.0.2=pyhd8ed1ab_0
asynctest=0.13.0=py_0
attrs=22.1.0=pyh71513ae_1
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py37h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blessings=1.7=py37h89c1867_1006
blinker=1.5=pyhd8ed1ab_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h540881e_1004
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.9.24=ha878542_0
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py37h74dc2b5_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=py37h89c1867_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=35.0.0=py37hf1a17b8_2
cudatoolkit=11.0.221=h6bb024c_0
cudnn=8.2.1.32=h86fa8c9_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py37h06a4308_0
expat=2.4.9=h6a678d5_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py37h7f8727e_0
future=0.18.2=py37h89c1867_5
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
google-auth=2.12.0=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
gpustat=0.6.0=pyhd8ed1ab_1
grpcio=1.42.0=py37hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.11.4=py37h89c1867_0
importlib_metadata=1.5.0=py37_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py37h06a4308_0
ipython=7.31.1=py37h06a4308_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py37h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.4.0=py37h06a4308_0
jupyter=1.0.0=py37h06a4308_8
jupyter_client=7.2.2=py37h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py37h06a4308_0
jupyter_server=1.18.1=py37h06a4308_0
jupyterlab=3.4.4=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.12.0=py37h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=py37h295c915_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=12.1.0=h8d9b700_16
libgomp=12.1.0=h8d9b700_16
liblapack=3.9.0=12_linux64_mkl
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.18.1=h780b84a_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
libzlib=1.2.12=h166bdaf_3
lz4-c=1.9.3=h295c915_1
magma=2.5.4=h4a2bb80_2
markdown=3.3.4=py37h06a4308_0
markupsafe=2.1.1=py37h540881e_1
matplotlib=3.5.2=py37h06a4308_0
matplotlib-base=3.5.2=py37hf590b9c_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37h14c3975_1001
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
multidict=6.0.2=py37h540881e_1
munkres=1.1.4=py_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py37h06a4308_0
nbconvert=6.4.4=py37h06a4308_0
nbformat=5.3.0=py37h06a4308_0
nccl=2.12.12.1=h17a0586_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py37h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
notebook=6.4.12=py37h06a4308_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
nvidia-ml=7.352.0=py_0
oauthlib=3.2.1=pyhd8ed1ab_0
openssl=1.1.1q=h166bdaf_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py37hace64e9_1
pip=22.2.2=py37h06a4308_0
ply=3.11=py37_0
prometheus_client=0.14.1=py37h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.18.1=py37hcd2ae1e_0
psutil=5.9.0=py37h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.5.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.15.7=py37h6a678d5_1
pyqt5-sip=12.11.0=py37h6a678d5_1
pyrsistent=0.18.0=py37heee7806_0
pysocks=1.7.1=py37h89c1867_5
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.9.1=cuda110py37h4a33b93_3
pytorch-gpu=1.9.1=cuda110py37h5b0ac8e_3
pytz=2022.1=py37h06a4308_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyzmq=23.2.0=py37h6a678d5_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtconsole=5.3.1=py37h06a4308_1
qtpy=2.0.1=pyhd3eb1b0_0
qtwebkit=5.212=h4eab89a_4
readline=8.1.2=h7f8727e_1
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=59.5.0=py37h89c1867_0
sip=6.6.2=py37h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h9b69904_2
sniffio=1.2.0=py37h06a4308_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.39.3=h5082296_0
tensorboard=2.10.1=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py37h38fbfac_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
terminado=0.13.1=py37h06a4308_0
testpath=0.6.0=py37h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.9.1=py37
torchvision=0.10.1=py37cuda110hbff3922_0_cuda
tornado=6.2=py37h5eee18b_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py37h06a4308_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
websocket-client=0.58.0=py37h06a4308_4
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py37h06a4308_0
xz=5.2.6=h5eee18b_0
yarl=1.7.2=py37h540881e_2
zeromq=4.3.4=h2531618_0
zipp=3.8.1=pyhd8ed1ab_0
zlib=1.2.12=h5eee18b_3
zstd=1.5.2=ha4553b6_0
